requirements.txt
google-genai
groq
pymupdf
pymupdf4llm
rarfile
libclang